+++ src/ftxui/component/menu.cpp
@@ -318,8 +318,7 @@ class MenuBase : public ComponentBase, public MenuOption {
 
       TakeFocus();
       focused_entry() = i;
-      if (event.mouse().button == Mouse::Left &&
-          event.mouse().motion == Mouse::Released) {
+      if (event.mouse().IsPressed()) {
         if (selected() != i) {
           selected() = i;
           selected_previous_ = selected();
@@ -683,8 +682,7 @@ Component MenuEntry(MenuEntryOption option) {
         return false;
       }
 
-      if (event.mouse().button == Mouse::Left &&
-          event.mouse().motion == Mouse::Released) {
+      if (event.mouse().IsPressed()) {
         TakeFocus();
         return true;
       }